Add files via upload
I started working on the project on 21st of March, and got problems working on it. I had studied and got experience in developing codes in HTML and CSS, however I didn’t know much about Javascript, which caused difficulties in understanding what to work on. Because of this, a big portion of complicated formulas were developed with the help of AI, and also used to check if everything is working correctly. Due to the complicated structure of certain elements such as filtering films, selecting tickets and seats, and recommending movies based on viewing history, I didn’t know what to do.
However parts such as  I managed to do myself:
I was able to plan out and do sketches for how the prototype should look and be structured.
While with the help from Bootstrap, I was able to structure the elements and connect the pages together with HTML using previous experiences in coding. As well as stylize the elements like buttons, and text with CSS.
Using information and resources that were given by AI and from sites that gave advice on coding, I was able to construct and change codes in order to function as it was intended.
I had been trying to work on it as much as possible, adding changes everyday and finally finished today on 28th March. Hopefully my efforts were enough.